Commit before I make a big change to spriting

+++ zephyrapp/game/src/cat_main.h
@@ -1,11 +1,13 @@
 #pragma once
 
+#include <stdint.h>
+
 extern int logged_sleep;
 
-extern int saved_version_major;
-extern int saved_version_minor;
-extern int saved_version_patch;
-extern int saved_version_push;
+extern uint8_t saved_version_major;
+extern uint8_t saved_version_minor;
+extern uint8_t saved_version_patch;
+extern uint8_t saved_version_push;
 
 #ifdef CAT_DESKTOP
 int CAT_load_sleep();
